Fix mismatching function declaration in header
Noticed this while trying to compile CSE2 with winelib.

+++ src/Sound.h
@@ -38,7 +38,7 @@ extern LPDIRECTSOUNDBUFFER lpSECONDARYBUFFER[SE_MAX];
 BOOL InitDirectSound(HWND hwnd);
 void EndDirectSound(void);
 void PlaySoundObject(int no, int mode);
-void ChangeSoundFrequency(int no, unsigned long rate);
+void ChangeSoundFrequency(int no, DWORD rate);
 void ChangeSoundVolume(int no, long volume);
 void ChangeSoundPan(int no, long pan);
 int MakePixToneObject(const PIXTONEPARAMETER *ptp, int ptp_num, int no);
